mcp-server/requirements.txt
# Core dependencies
mcp-python==0.1.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# Document processing
langchain>=0.1.0
langchain_community>=0.0.1
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
docx2txt>=0.8
python-magic>=0.4.27
unstructured>=0.10.0
pymupdf>=1.23.0

# Vector store
qdrant-client>=1.6.0
langchain-qdrant>=0.0.1

# Utilities
numpy>=1.24.0
python-dateutil>=2.8.2
structlog>=23.1.0

# Missing dependency
langchain_huggingface
